+++ sdk/communication/Azure.Communication.Chat/api/Azure.Communication.Chat.netstandard2.0.cs
@@ -44,16 +44,25 @@ namespace Azure.Communication.Chat
     public partial class ChatMessage
     {
         internal ChatMessage() { }
-        public string Content { get { throw null; } }
-        public System.DateTimeOffset? CreatedOn { get { throw null; } }
+        public Azure.Communication.Chat.ChatMessageContent Content { get { throw null; } }
+        public System.DateTimeOffset CreatedOn { get { throw null; } }
         public System.DateTimeOffset? DeletedOn { get { throw null; } }
         public System.DateTimeOffset? EditedOn { get { throw null; } }
         public string Id { get { throw null; } }
-        public Azure.Communication.Chat.ChatMessagePriority? Priority { get { throw null; } }
+        public Azure.Communication.Chat.ChatMessagePriority Priority { get { throw null; } }
         public Azure.Communication.CommunicationUserIdentifier Sender { get { throw null; } }
         public string SenderDisplayName { get { throw null; } }
-        public string Type { get { throw null; } }
+        public string SequenceId { get { throw null; } }
+        public Azure.Communication.Chat.ChatMessageType Type { get { throw null; } }
         public string Version { get { throw null; } }
+    }
+    public partial class ChatMessageContent
+    {
+        internal ChatMessageContent() { }
+        public Azure.Communication.CommunicationUserIdentifier? Initiator { get { throw null; } }
+        public string Message { get { throw null; } }
+        public System.Collections.Generic.IReadOnlyList<Azure.Communication.Chat.ChatParticipant> Participants { get { throw null; } }
+        public string Topic { get { throw null; } }
     }
     [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
     public readonly partial struct ChatMessagePriority : System.IEquatable<Azure.Communication.Chat.ChatMessagePriority>
@@ -77,15 +86,36 @@ namespace Azure.Communication.Chat
     {
         internal ChatMessageReadReceipt() { }
         public string ChatMessageId { get { throw null; } }
-        public System.DateTimeOffset? ReadOn { get { throw null; } }
+        public System.DateTimeOffset ReadOn { get { throw null; } }
         public Azure.Communication.CommunicationUserIdentifier Sender { get { throw null; } }
+    }
+    [System.Runtime.InteropServices.StructLayoutAttribute(System.Runtime.InteropServices.LayoutKind.Sequential)]
+    public readonly partial struct ChatMessageType : System.IEquatable<Azure.Communication.Chat.ChatMessageType>
+    {
+        private readonly object _dummy;
+        private readonly int _dummyPrimitive;
+        public ChatMessageType(string value) { throw null; }
+        public static Azure.Communication.Chat.ChatMessageType Html { get { throw null; } }
+        public static Azure.Communication.Chat.ChatMessageType ParticipantAdded { get { throw null; } }
+        public static Azure.Communication.Chat.ChatMessageType ParticipantRemoved { get { throw null; } }
+        public static Azure.Communication.Chat.ChatMessageType Text { get { throw null; } }
+        public static Azure.Communication.Chat.ChatMessageType TopicUpdated { get { throw null; } }
+        public bool Equals(Azure.Communication.Chat.ChatMessageType other) { throw null; }
+        [System.ComponentModel.EditorBrowsableAttribute(System.ComponentModel.EditorBrowsableState.Never)]
+        public override bool Equals(object obj) { throw null; }
+        [System.ComponentModel.EditorBrowsableAttribute(System.ComponentModel.EditorBrowsableState.Never)]
+        public override int GetHashCode() { throw null; }
+        public static bool operator ==(Azure.Communication.Chat.ChatMessageType left, Azure.Communication.Chat.ChatMessageType right) { throw null; }
+        public static implicit operator Azure.Communication.Chat.ChatMessageType (string value) { throw null; }
+        public static bool operator !=(Azure.Communication.Chat.ChatMessageType left, Azure.Communication.Chat.ChatMessageType right) { throw null; }
+        public override string ToString() { throw null; }
     }
     public static partial class ChatModelFactory
     {
         [System.ComponentModel.EditorBrowsableAttribute(System.ComponentModel.EditorBrowsableState.Never)]
-        public static Azure.Communication.Chat.ChatMessage ChatMessage(string id, string type, Azure.Communication.Chat.ChatMessagePriority? priority, string version, string content, string senderDisplayName, System.DateTimeOffset? createdOn, string senderId, System.DateTimeOffset? deletedOn, System.DateTimeOffset? editedOn) { throw null; }
+        public static Azure.Communication.Chat.ChatMessage ChatMessage(string id, Azure.Communication.Chat.ChatMessageType type, Azure.Communication.Chat.ChatMessagePriority priority, string sequenceId, string version, Azure.Communication.Chat.ChatMessageContent content, string senderDisplayName, System.DateTimeOffset createdOn, string senderId, System.DateTimeOffset? deletedOn, System.DateTimeOffset? editedOn) { throw null; }
         [System.ComponentModel.EditorBrowsableAttribute(System.ComponentModel.EditorBrowsableState.Never)]
-        public static Azure.Communication.Chat.ChatMessageReadReceipt ChatMessageReadReceipt(string senderId, string chatMessageId, System.DateTimeOffset? readOn) { throw null; }
+        public static Azure.Communication.Chat.ChatMessageReadReceipt ChatMessageReadReceipt(string senderId, string chatMessageId, System.DateTimeOffset readOn) { throw null; }
         [System.ComponentModel.EditorBrowsableAttribute(System.ComponentModel.EditorBrowsableState.Never)]
         public static Azure.Communication.Chat.ChatThreadInfo ChatThreadInfo(string id, string topic, System.DateTimeOffset? deletedOn, System.DateTimeOffset? lastMessageReceivedOn) { throw null; }
     }
@@ -93,14 +123,14 @@ namespace Azure.Communication.Chat
     {
         public ChatParticipant(Azure.Communication.CommunicationUserIdentifier communicationUserIdentifier) { }
         public string? DisplayName { get { throw null; } set { } }
-        public System.DateTimeOffset? ShareHistoryTime { get { throw null; } set { } }
+        public System.DateTimeOffset ShareHistoryTime { get { throw null; } set { } }
         public Azure.Communication.CommunicationUserIdentifier User { get { throw null; } set { } }
     }
     public partial class ChatThread
     {
         internal ChatThread() { }
         public Azure.Communication.CommunicationUserIdentifier CreatedBy { get { throw null; } }
-        public System.DateTimeOffset? CreatedOn { get { throw null; } }
+        public System.DateTimeOffset CreatedOn { get { throw null; } }
         public System.DateTimeOffset? DeletedOn { get { throw null; } }
         public string Id { get { throw null; } }
         public string Topic { get { throw null; } }